requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.21.0
annotated-types==0.7.0
apiclient==1.0.4
APScheduler==3.10.4
attrs==23.2.0
beautifulsoup4==4.13.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
environ==1.0
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.1.1
gspread==6.1.2
httplib2==0.22.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
pytz==2024.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
YooMoney==0.1.0
